added float cast to avoid numbers below 1 causing output to be 0

+++ utils/Algorithms/Algorithms.cpp
@@ -15,7 +15,7 @@ int invertIntInRange(int min, int max, int num)
 
 int scaleIntToRange(int input, int minInput, int maxInput, int minOutput, int maxOutput)
 {
-    return ( (maxOutput - minOutput) / (maxInput - minInput) ) * (input - minInput);
+    return ((float)(maxOutput - minOutput) / (maxInput - minInput)) * (input - minInput);
 }
 
 /**
